Emit [u]int_[32|64]_t for long and long long IDL types
Removed Serializer specializations for `long long` and
`unsigned long long`, which fixes the MSVC build where these types are
typedefs of int64_t and uint64_t respectively.

+++ src/bindings/gpurenderpipeline.cc
@@ -16,7 +16,7 @@ GPURenderPipeline::GPURenderPipeline(wgpu::RenderPipeline pipeline)
     : pipeline_(pipeline) {}
 
 interop::Interface<interop::GPUBindGroupLayout>
-GPURenderPipeline::getBindGroupLayout(Napi::Env env, unsigned long index) {
+GPURenderPipeline::getBindGroupLayout(Napi::Env env, uint32_t index) {
   return interop::GPUBindGroupLayout::Create<GPUBindGroupLayout>(
       env, pipeline_.GetBindGroupLayout(index));
 }
